chore: rename file name for error modal
