+++ lib/app/features/wallets/views/pages/coins_flow/contact_picker/contact_picker_modal.dart
@@ -7,7 +7,7 @@ import 'package:ion/app/extensions/extensions.dart';
 import 'package:ion/app/features/user/model/user_metadata.c.dart';
 import 'package:ion/app/features/user/pages/user_picker_sheet/user_picker_sheet.dart';
 import 'package:ion/app/features/wallets/providers/networks_provider.c.dart';
-import 'package:ion/app/features/wallets/views/pages/contact_without_wallet_error.dart';
+import 'package:ion/app/features/wallets/views/pages/contact_without_wallet_error_modal.dart';
 import 'package:ion/app/router/components/navigation_app_bar/navigation_app_bar.dart';
 import 'package:ion/app/router/components/navigation_app_bar/navigation_close_button.dart';
 import 'package:ion/app/router/components/sheet_content/sheet_content.dart';
