+++ lib/app/features/wallets/views/pages/contact_without_wallet_error_modal.dart
